Move global id generation to IntegralLines.h and remove getVertexLocalIdIfExists
