+++ core/base/abstractTriangulation/AbstractTriangulation.h
@@ -2663,27 +2663,6 @@ namespace ttk {
       return this->getVertexLocalIdInternal(gvid);
     }
 
-    virtual inline SimplexId
-      getVertexLocalIdIfExists(const SimplexId gvid) const {
-      if(!isRunningWithMPI()) {
-        return gvid;
-      }
-#ifndef TTK_ENABLE_KAMIKAZE
-      if(this->getDimensionality() != 1 && this->getDimensionality() != 2
-         && this->getDimensionality() != 3) {
-        this->printErr("Only 1D, 2D and 3D datasets are supported");
-        return -1;
-      }
-      if(!this->hasPreconditionedDistributedVertices_) {
-        this->printErr("VertexLocalId query without pre-process!");
-        this->printErr(
-          "Please call preconditionDistributedVertices() in a pre-process.");
-        return -1;
-      }
-#endif // TTK_ENABLE_KAMIKAZE
-      return this->getVertexLocalIdIfExistsInternal(gvid);
-    }
-
     virtual inline SimplexId getCellGlobalId(const SimplexId lcid) const {
 #ifndef TTK_ENABLE_KAMIKAZE
       if(this->getDimensionality() != 1 && this->getDimensionality() != 2
@@ -2956,11 +2935,6 @@ namespace ttk {
 
     virtual inline SimplexId
       getVertexLocalIdInternal(const SimplexId ttkNotUsed(gvid)) const {
-      return -1;
-    }
-
-    virtual inline SimplexId
-      getVertexLocalIdIfExistsInternal(const SimplexId ttkNotUsed(gvid)) const {
       return -1;
     }
