+++ core/base/integralLines/IntegralLines.h
@@ -32,39 +32,69 @@
 #endif
 
 namespace ttk {
-
-  struct IntegralLine {
-    std::vector<ttk::SimplexId> *trajectory;
-    std::vector<double> *distanceFromSeed;
-    std::vector<ttk::SimplexId> *localVertexIdentifier;
-    ttk::SimplexId seedIdentifier;
-    ttk::SimplexId forkIdentifier = -1;
-  };
+  namespace intgl {
+    struct IntegralLine {
+      std::vector<ttk::SimplexId> *trajectory;
+      std::vector<double> *distanceFromSeed;
+      std::vector<ttk::SimplexId> *localVertexIdentifier;
+      ttk::SimplexId seedIdentifier;
+      ttk::SimplexId forkIdentifier = -1;
+    };
 
 #ifdef TTK_ENABLE_MPI
-  /* Counters for keeping tracks of number of integral lines
-   * left to compute. Each counter is aligned on a cache line
-   * to prevent false-sharing
-   */
-  static int finishedElement_ __attribute__((aligned(64)));
-  static int addedElement_ __attribute__((aligned(64)));
+    /**
+     * @brief Struct used for sorting ghost data during the generation of global
+     * ids. A GhostElementsToSort object is created for each segment of integral
+     * line that ends or begins with a ghost vertex. minLocalVertexId,
+     * seedIdentifier and forkIdentifier are used to identify this segment of
+     * integral line on different processes.
+     */
+    struct GhostElementsToSort {
+      ttk::SimplexId ownedGlobalId;
+      ttk::SimplexId minLocalVertexId;
+      ttk::SimplexId seedIdentifier;
+      ttk::SimplexId forkIdentifier{-1};
+      ttk::SimplexId globalEdgeId{-1};
+      ttk::SimplexId ghostVertexLocalId;
+      ttk::SimplexId ghostEdgeLocalId;
+    };
 
-  /*
-   * For each integral line continuing on another process, we send one layer of
-   * ghost cells and we keep one layers for this process, meaning that vertices
-   * Id1 belongs to this process and Id2 belongs to a neighboring
-   * process.
-   */
-  struct ElementToBeSent {
-    ttk::SimplexId Id1;
-    ttk::SimplexId Id2;
-    double DistanceFromSeed1;
-    double DistanceFromSeed2;
-    ttk::SimplexId LocalVertexIdentifier1;
-    ttk::SimplexId LocalVertexIdentifier2;
-    ttk::SimplexId SeedIdentifier;
-    ttk::SimplexId ForkIdentifier;
-  };
+    inline bool operator<(const GhostElementsToSort &left,
+                          const GhostElementsToSort &right) {
+      if(left.seedIdentifier != right.seedIdentifier) {
+        return left.seedIdentifier < right.seedIdentifier;
+      }
+      if(left.forkIdentifier != right.forkIdentifier) {
+        return left.forkIdentifier < right.forkIdentifier;
+      }
+      return left.minLocalVertexId < right.minLocalVertexId;
+    };
+
+    /* Counters for keeping tracks of number of integral lines
+     * left to compute. Each counter is aligned on a cache line
+     * to prevent false-sharing
+     */
+    static int finishedElement_ __attribute__((aligned(64)));
+    static int addedElement_ __attribute__((aligned(64)));
+
+    /*
+     * For each integral line continuing on another process, we send one layer
+     * of ghost cells and we keep one layers for this process, meaning that
+     * vertices Id1 belongs to this process and Id2 belongs to a neighboring
+     * process.
+     */
+    struct ElementToBeSent {
+      ttk::SimplexId Id1;
+      ttk::SimplexId Id2;
+      double DistanceFromSeed1;
+      double DistanceFromSeed2;
+      ttk::SimplexId LocalVertexIdentifier1;
+      ttk::SimplexId LocalVertexIdentifier2;
+      ttk::SimplexId SeedIdentifier;
+      ttk::SimplexId ForkIdentifier;
+    };
+  } // namespace intgl
+
 #endif
 
   enum Direction { Forward = 0, Backward };
@@ -89,7 +119,7 @@ namespace ttk {
      */
     template <class triangulationType = ttk::AbstractTriangulation>
     void computeIntegralLine(const triangulationType *triangulation,
-                             ttk::IntegralLine integralLine,
+                             ttk::intgl::IntegralLine integralLine,
                              const ttk::SimplexId *offsets) const;
 
     /**
@@ -104,7 +134,7 @@ namespace ttk {
      */
     template <class triangulationType>
     void createTask(const triangulationType *triangulation,
-                    std::vector<ttk::IntegralLine> &chunkIntegralLine,
+                    std::vector<ttk::intgl::IntegralLine> &chunkIntegralLine,
                     const ttk::SimplexId *offsets,
                     int nbElement) const;
     /**
@@ -122,11 +152,12 @@ namespace ttk {
      * @param seeds starting points of the integral lines to be computed
      */
     template <class triangulationType>
-    void prepareForTask(const triangulationType *triangulation,
-                        std::vector<ttk::IntegralLine> &chunkIntegralLine,
-                        int startingIndex,
-                        int nbElement,
-                        std::vector<SimplexId> *seeds) const;
+    void
+      prepareForTask(const triangulationType *triangulation,
+                     std::vector<ttk::intgl::IntegralLine> &chunkIntegralLine,
+                     int startingIndex,
+                     int nbElement,
+                     std::vector<SimplexId> *seeds) const;
 
     inline void setVertexNumber(const SimplexId &vertexNumber) {
       vertexNumber_ = vertexNumber;
@@ -148,6 +179,53 @@ namespace ttk {
 #ifdef TTK_ENABLE_MPI
 
     /**
+     * @brief Constructs the global identifiers for vertices and edges.
+     *
+     * @tparam triangulationType
+     * @param globalVertexId vector of global identifiers for vertices
+     * @param globalCellId vector of global identifiers for cells
+     * @param seedIdentifiers linked list of identifiers of seed, one for each
+     * integral line
+     * @param forkIdentifiers linked list of identifiers of forks, one for each
+     * integral line
+     * @param trajectories linked list of identifiers of forks, one for each
+     * integral line
+     * @param localVertexIdentifiers linked list of identifiers, one for each
+     * vertex of an integral lines. It represents the number of the vertex in
+     * the integral line.
+     * @param triangulation Triangulation
+     * @return int
+     */
+    template <typename triangulationType>
+    int getGlobalIdentifiers(
+      std::vector<ttk::SimplexId> &globalVertexId,
+      std::vector<ttk::SimplexId> &globalCellId,
+      std::vector<ttk::ArrayLinkedList<ttk::SimplexId, TABULAR_SIZE>>
+        &seedIdentifiers,
+      std::vector<ttk::ArrayLinkedList<ttk::SimplexId, TABULAR_SIZE>>
+        &forkIdentifiers,
+      std::vector<ttk::ArrayLinkedList<std::vector<ttk::SimplexId>,
+                                       TABULAR_SIZE>> &trajectories,
+      std::vector<ttk::ArrayLinkedList<std::vector<ttk::SimplexId>,
+                                       TABULAR_SIZE>> &localVertexIdentifiers,
+      triangulationType *triangulation);
+
+    /**
+     * @brief Sorts ghosts and exchanges their global identifiers between
+     * processes.
+     *
+     * @param unmatchedGhosts Ghosts vertices and edges for which the global
+     * identifier is to be determined
+     * @param globalVertexId vector of global identifiers for vertices
+     * @param globalCellId vector of global identifiers for cells
+     * @return int 0 for success
+     */
+    int exchangeGhosts(std::vector<std::vector<ttk::intgl::GhostElementsToSort>>
+                         &unmatchedGhosts,
+                       std::vector<ttk::SimplexId> &globalVertexId,
+                       std::vector<ttk::SimplexId> &globalCellId);
+
+    /**
      * @brief Checks if an integral line should be continued on another process
      * or not. If so, encapsulate the necessary data in a struct and stores it
      * in toSend_.
@@ -160,7 +238,7 @@ namespace ttk {
      */
     template <class triangulationType>
     void storeToSendIfNecessary(const triangulationType *triangulation,
-                                ttk::IntegralLine integralLine,
+                                ttk::intgl::IntegralLine integralLine,
                                 bool &isMax) const;
 
     /**
@@ -179,15 +257,16 @@ namespace ttk {
      * @param offsets Order array of the scalar array
      */
     template <class triangulationType>
-    void receiveElement(const triangulationType *triangulation,
-                        ElementToBeSent &element,
-                        std::vector<ttk::IntegralLine> &chunkIntegralLine,
-                        int &index,
-                        int taskSize,
-                        const ttk::SimplexId *offsets);
+    void
+      receiveElement(const triangulationType *triangulation,
+                     intgl::ElementToBeSent &element,
+                     std::vector<ttk::intgl::IntegralLine> &chunkIntegralLine,
+                     int &index,
+                     int taskSize,
+                     const ttk::SimplexId *offsets);
 
-    inline void
-      setToSend(std::vector<std::vector<std::vector<ElementToBeSent>>> *send) {
+    inline void setToSend(
+      std::vector<std::vector<std::vector<intgl::ElementToBeSent>>> *send) {
       toSend_ = send;
     }
 
@@ -216,14 +295,14 @@ namespace ttk {
            getMPIType(id), getMPIType(id), getMPIType(id), getMPIType(id)};
       int lengths[] = {1, 1, 1, 1, 1, 1, 1, 1};
       const long int mpi_offsets[]
-        = {offsetof(ElementToBeSent, Id1),
-           offsetof(ElementToBeSent, Id2),
-           offsetof(ElementToBeSent, DistanceFromSeed1),
-           offsetof(ElementToBeSent, DistanceFromSeed2),
-           offsetof(ElementToBeSent, LocalVertexIdentifier1),
-           offsetof(ElementToBeSent, LocalVertexIdentifier2),
-           offsetof(ElementToBeSent, SeedIdentifier),
-           offsetof(ElementToBeSent, ForkIdentifier)};
+        = {offsetof(intgl::ElementToBeSent, Id1),
+           offsetof(intgl::ElementToBeSent, Id2),
+           offsetof(intgl::ElementToBeSent, DistanceFromSeed1),
+           offsetof(intgl::ElementToBeSent, DistanceFromSeed2),
+           offsetof(intgl::ElementToBeSent, LocalVertexIdentifier1),
+           offsetof(intgl::ElementToBeSent, LocalVertexIdentifier2),
+           offsetof(intgl::ElementToBeSent, SeedIdentifier),
+           offsetof(intgl::ElementToBeSent, ForkIdentifier)};
       MPI_Type_create_struct(
         8, lengths, mpi_offsets, types, &(this->MessageType));
       MPI_Type_commit(&(this->MessageType));
@@ -319,7 +398,8 @@ namespace ttk {
     ttk::ScalarFieldCriticalPoints scalarFieldCriticalPoints_;
 
 #ifdef TTK_ENABLE_MPI
-    std::vector<std::vector<std::vector<ElementToBeSent>>> *toSend_{nullptr};
+    std::vector<std::vector<std::vector<intgl::ElementToBeSent>>> *toSend_{
+      nullptr};
     int neighborNumber_;
     std::unordered_map<int, int> neighborsToId_;
     std::vector<int> neighbors_;
@@ -335,12 +415,12 @@ namespace ttk {
 template <class triangulationType>
 void ttk::IntegralLines::receiveElement(
   const triangulationType *triangulation,
-  ElementToBeSent &element,
-  std::vector<ttk::IntegralLine> &chunkIntegralLine,
+  intgl::ElementToBeSent &element,
+  std::vector<ttk::intgl::IntegralLine> &chunkIntegralLine,
   int &index,
   int taskSize,
   const ttk::SimplexId *offsets) {
-  ttk::IntegralLine integralLine = ttk::IntegralLine{
+  ttk::intgl::IntegralLine integralLine = ttk::intgl::IntegralLine{
     nullptr, nullptr, nullptr, element.SeedIdentifier, element.ForkIdentifier};
   // Create integral line object on this process
   int threadNum{0};
@@ -383,7 +463,7 @@ void ttk::IntegralLines::receiveElement(
 template <class triangulationType>
 void ttk::IntegralLines::storeToSendIfNecessary(
   const triangulationType *triangulation,
-  ttk::IntegralLine integralLine,
+  ttk::intgl::IntegralLine integralLine,
   bool &isMax) const {
 #ifdef TTK_ENABLE_MPI
   if(ttk::isRunningWithMPI()) {
@@ -392,14 +472,15 @@ void ttk::IntegralLines::storeToSendIfNecessary(
       int rankArray
         = triangulation->getVertexRank(integralLine.trajectory->back());
       if(rankArray != ttk::MPIrank_) {
-        ElementToBeSent element = ElementToBeSent{-1,
-                                                  -1,
-                                                  0,
-                                                  0,
-                                                  -1,
-                                                  -1,
-                                                  integralLine.seedIdentifier,
-                                                  integralLine.forkIdentifier};
+        intgl::ElementToBeSent element
+          = intgl::ElementToBeSent{-1,
+                                   -1,
+                                   0,
+                                   0,
+                                   -1,
+                                   -1,
+                                   integralLine.seedIdentifier,
+                                   integralLine.forkIdentifier};
         element.Id2
           = triangulation->getVertexGlobalId(integralLine.trajectory->back());
         element.Id1 = triangulation->getVertexGlobalId(
@@ -450,7 +531,7 @@ inline void
 template <class triangulationType>
 void ttk::IntegralLines::computeIntegralLine(
   const triangulationType *triangulation,
-  ttk::IntegralLine integralLine,
+  ttk::intgl::IntegralLine integralLine,
   const SimplexId *offsets) const {
   double distance = integralLine.distanceFromSeed->back();
   ttk::SimplexId v = integralLine.trajectory->back();
@@ -478,7 +559,7 @@ void ttk::IntegralLines::computeIntegralLine(
 #if TTK_ENABLE_OPENMP
 #pragma omp atomic update seq_cst
 #endif
-        finishedElement_++;
+        ttk::intgl::finishedElement_++;
       } else {
         this->storeToSendIfNecessary<triangulationType>(
           triangulation, integralLine, isMax);
@@ -501,11 +582,11 @@ void ttk::IntegralLines::computeIntegralLine(
 #if TTK_ENABLE_OPENMP
 #pragma omp atomic update seq_cst
 #endif
-        finishedElement_++;
+        ttk::intgl::finishedElement_++;
 #if TTK_ENABLE_OPENMP
 #pragma omp atomic update seq_cst
 #endif
-        addedElement_ += numberOfComponents;
+        ttk::intgl::addedElement_ += numberOfComponents;
 #endif
         isMax = true;
         for(int i = 0; i < numberOfComponents; i++) {
@@ -519,9 +600,9 @@ void ttk::IntegralLines::computeIntegralLine(
 #else
           ttk::SimplexId forkIdentifier = vnext;
 #endif
-          ttk::IntegralLine integralLineFork
-            = ttk::IntegralLine{nullptr, nullptr, nullptr,
-                                integralLine.seedIdentifier, forkIdentifier};
+          ttk::intgl::IntegralLine integralLineFork = ttk::intgl::IntegralLine{
+            nullptr, nullptr, nullptr, integralLine.seedIdentifier,
+            forkIdentifier};
           triangulation->getVertexPoint(vnext, p1[0], p1[1], p1[2]);
           double distanceFork = Geometry::distance(p0, p1, 3);
           int threadNum{0};
@@ -603,7 +684,7 @@ void ttk::IntegralLines::prepareForTask(
 #else
   const triangulationType *ttkNotUsed(triangulation),
 #endif
-  std::vector<ttk::IntegralLine> &chunkIntegralLine,
+  std::vector<ttk::intgl::IntegralLine> &chunkIntegralLine,
   int startingIndex,
   int nbElement,
   std::vector<SimplexId> *seeds) const {
@@ -638,7 +719,7 @@ void ttk::IntegralLines::prepareForTask(
 template <class triangulationType>
 void ttk::IntegralLines::createTask(
   const triangulationType *triangulation,
-  std::vector<ttk::IntegralLine> &chunkIntegralLine,
+  std::vector<ttk::intgl::IntegralLine> &chunkIntegralLine,
   const ttk::SimplexId *offsets,
   int nbElement) const {
 #if TTK_ENABLE_OPENMP
@@ -659,18 +740,19 @@ int ttk::IntegralLines::execute(triangulationType *triangulation) {
 
 #ifdef TTK_ENABLE_MPI
   keepWorking_ = 1;
-  finishedElement_ = 0;
-  addedElement_ = 0;
+  ttk::intgl::finishedElement_ = 0;
+  ttk::intgl::addedElement_ = 0;
 #endif
   const SimplexId *offsets = inputOffsets_;
   std::vector<SimplexId> *seeds = vertexIdentifierScalarField_;
   Timer t;
 
-  std::vector<ttk::IntegralLine> chunkIntegralLine(chunkSize_);
+  std::vector<ttk::intgl::IntegralLine> chunkIntegralLine(chunkSize_);
   int taskNumber = (int)seedNumber_ / chunkSize_;
 #ifdef TTK_ENABLE_OPENMP
 #ifdef TTK_ENABLE_MPI
-#pragma omp parallel shared(finishedElement_, toSend_, addedElement_) \
+#pragma omp parallel shared(                                        \
+  ttk::intgl::finishedElement_, toSend_, ttk::intgl::addedElement_) \
   num_threads(threadNumber_)
   {
 #else
@@ -704,8 +786,8 @@ int ttk::IntegralLines::execute(triangulationType *triangulation) {
     int finishedElementReceived = 0;
     std::vector<int> sendMessageSize(neighborNumber_);
     std::vector<int> recvMessageSize(neighborNumber_);
-    std::vector<std::vector<ElementToBeSent>> send_buf(neighborNumber_);
-    std::vector<std::vector<ElementToBeSent>> recv_buf(neighborNumber_);
+    std::vector<std::vector<intgl::ElementToBeSent>> send_buf(neighborNumber_);
+    std::vector<std::vector<intgl::ElementToBeSent>> recv_buf(neighborNumber_);
     for(i = 0; i < neighborNumber_; i++) {
       send_buf.reserve((int)seedNumber_ * 0.005);
       recv_buf.reserve((int)seedNumber_ * 0.005);
@@ -716,12 +798,12 @@ int ttk::IntegralLines::execute(triangulationType *triangulation) {
     int index;
     int totalMessageSize;
     while(keepWorking_) {
-      finishedElement_ -= addedElement_;
-      addedElement_ = 0;
+      ttk::intgl::finishedElement_ -= ttk::intgl::addedElement_;
+      ttk::intgl::addedElement_ = 0;
       // Exchange of the number of integral lines finished on all processes
-      MPI_Allreduce(&finishedElement_, &finishedElementReceived, 1, MPI_INTEGER,
-                    MPI_SUM, ttk::MPIcomm_);
-      finishedElement_ = 0;
+      MPI_Allreduce(&ttk::intgl::finishedElement_, &finishedElementReceived, 1,
+                    MPI_INTEGER, MPI_SUM, ttk::MPIcomm_);
+      ttk::intgl::finishedElement_ = 0;
       // Update the number of integral lines left to compute
       globalElementCounter_ -= finishedElementReceived;
       // Stop working in case there are no more computation to be done
@@ -769,7 +851,7 @@ int ttk::IntegralLines::execute(triangulationType *triangulation) {
         }
         // Extraction of the received data and creation of the tasks
 #ifdef TTK_ENABLE_OPENMP
-#pragma omp parallel shared(finishedElement_, toSend_) \
+#pragma omp parallel shared(ttk::intgl::finishedElement_, toSend_) \
   num_threads(threadNumber_)
         {
 #pragma omp master
@@ -807,3 +889,241 @@ int ttk::IntegralLines::execute(triangulationType *triangulation) {
   }
   return 0;
 }
+
+#ifdef TTK_ENABLE_MPI
+template <typename triangulationType>
+int ttk::IntegralLines::getGlobalIdentifiers(
+  std::vector<ttk::SimplexId> &globalVertexId,
+  std::vector<ttk::SimplexId> &globalCellId,
+  std::vector<ttk::ArrayLinkedList<ttk::SimplexId, TABULAR_SIZE>>
+    &seedIdentifiers,
+  std::vector<ttk::ArrayLinkedList<ttk::SimplexId, TABULAR_SIZE>>
+    &forkIdentifiers,
+  std::vector<ttk::ArrayLinkedList<std::vector<ttk::SimplexId>, TABULAR_SIZE>>
+    &trajectories,
+  std::vector<ttk::ArrayLinkedList<std::vector<ttk::SimplexId>, TABULAR_SIZE>>
+    &localVertexIdentifiers,
+  triangulationType *triangulation) {
+  ttk::SimplexId outputVertexNumber = 0;
+  ttk::SimplexId outputCellNumber = 0;
+  ttk::SimplexId realVertexNumber = 0;
+  ttk::SimplexId realCellNumber = 0;
+  ttk::SimplexId intervalSize;
+  // Counts vertices and edges number (with and without ghosts)
+#ifdef TTK_ENABLE_OPENMP
+#pragma omp parallel for reduction(+:outputVertexNumber,outputCellNumber,realCellNumber,realVertexNumber) schedule(static,1) private(intervalSize)
+#endif
+  for(int thread = 0; thread < threadNumber_; thread++) {
+    std::list<std::array<std::vector<ttk::SimplexId>, TABULAR_SIZE>>::iterator
+      trajectory
+      = trajectories[thread].list.begin();
+    while(trajectory != trajectories[thread].list.end()) {
+      for(int i = 0; i < TABULAR_SIZE; i++) {
+        if((*trajectory)[i].size() > 0) {
+          intervalSize = static_cast<ttk::SimplexId>((*trajectory)[i].size());
+          outputVertexNumber += intervalSize;
+          outputCellNumber += intervalSize - 1;
+          if(triangulation->getVertexRank((*trajectory)[i].at(0))
+             != ttk::MPIrank_) {
+            intervalSize--;
+          }
+          if((*trajectory)[i].size() > 1) {
+            realCellNumber += intervalSize - 1;
+            if(triangulation->getVertexRank((*trajectory)[i].back())
+               != ttk::MPIrank_) {
+              intervalSize--;
+            }
+          }
+          realVertexNumber += intervalSize;
+        } else {
+          break;
+        }
+      }
+      trajectory++;
+    }
+  }
+
+  ttk::SimplexId vertIndex;
+  ttk::SimplexId cellIndex;
+
+  // Perform exclusive prefix sum to find local offset for vertices and
+  // cells
+  MPI_Exscan(&realVertexNumber, &vertIndex, 1, ttk::getMPIType(vertIndex),
+             MPI_SUM, ttk::MPIcomm_);
+  MPI_Exscan(&realCellNumber, &cellIndex, 1, ttk::getMPIType(cellIndex),
+             MPI_SUM, ttk::MPIcomm_);
+
+  // Rank 0 received garbage values, it is replaced by the correct offset
+  // (always 0)
+  if(ttk::MPIrank_ == 0) {
+    vertIndex = 0;
+    cellIndex = 0;
+  }
+
+  // Generate Global identifers and package ghost data for the process that owns
+  // the ghost data
+  ttk::SimplexId startCellId = 0;
+  ttk::SimplexId startVertexId = 0;
+  globalVertexId.resize(outputVertexNumber);
+  globalCellId.resize(outputCellNumber);
+  std::vector<std::vector<ttk::intgl::GhostElementsToSort>> unmatchedGhosts(
+    neighborNumber_);
+  std::vector<std::vector<ttk::SimplexId>> unmatchedGhostsVertexLocalId(
+    neighborNumber_);
+  std::vector<std::vector<ttk::SimplexId>> unmatchedGhostsEdgeLocalId(
+    neighborNumber_);
+  for(int thread = 0; thread < threadNumber_; thread++) {
+    std::list<std::array<ttk::SimplexId, TABULAR_SIZE>>::iterator forkIdentifier
+      = forkIdentifiers[thread].list.begin();
+    std::list<std::array<std::vector<ttk::SimplexId>, TABULAR_SIZE>>::iterator
+      localVertexIdentifier
+      = localVertexIdentifiers[thread].list.begin();
+    std::list<std::array<std::vector<ttk::SimplexId>, TABULAR_SIZE>>::iterator
+      trajectory
+      = trajectories[thread].list.begin();
+    std::list<std::array<ttk::SimplexId, TABULAR_SIZE>>::iterator seedIdentifier
+      = seedIdentifiers[thread].list.begin();
+    while(trajectory != trajectories[thread].list.end()) {
+      for(int i = 0; i < TABULAR_SIZE; i++) {
+        if((*trajectory)[i].size() > 0) {
+          if(triangulation->getVertexRank((*trajectory)[i].at(0))
+             != ttk::MPIrank_) {
+            globalVertexId.at(startVertexId) = -1;
+          } else {
+            globalVertexId.at(startVertexId) = vertIndex;
+            vertIndex++;
+          }
+          startVertexId++;
+          if((*trajectory)[i].size() > 1) {
+            if(triangulation->getVertexRank((*trajectory)[i].at(0))
+               != ttk::MPIrank_) {
+              globalCellId.at(startCellId) = -1;
+              unmatchedGhosts
+                .at(neighborsToId_[triangulation->getVertexRank(
+                  (*trajectory)[i].at(0))])
+                .push_back(ttk::intgl::GhostElementsToSort{
+                  vertIndex, (*localVertexIdentifier)[i].at(0),
+                  (*seedIdentifier)[i], (*forkIdentifier)[i], -1,
+                  startVertexId - 1, startCellId});
+            } else {
+              globalCellId.at(startCellId) = cellIndex;
+              cellIndex++;
+            }
+            startCellId++;
+            if((*trajectory)[i].size() > 2) {
+              std::iota(globalCellId.begin() + startCellId,
+                        globalCellId.begin() + startCellId
+                          + (*trajectory)[i].size() - 2,
+                        cellIndex);
+              std::iota(globalVertexId.begin() + startVertexId,
+                        globalVertexId.begin() + startVertexId
+                          + (*trajectory)[i].size() - 2,
+                        vertIndex);
+              startCellId += (*trajectory)[i].size() - 2;
+              startVertexId += (*trajectory)[i].size() - 2;
+              vertIndex += (*trajectory)[i].size() - 2;
+              cellIndex += (*trajectory)[i].size() - 2;
+            }
+            if(triangulation->getVertexRank((*trajectory)[i].back())
+               != ttk::MPIrank_) {
+              globalVertexId.at(startVertexId) = -1;
+              unmatchedGhosts
+                .at(neighborsToId_[triangulation->getVertexRank(
+                  (*trajectory)[i].back())])
+                .push_back(ttk::intgl::GhostElementsToSort{
+                  globalVertexId.at(startVertexId - 1),
+                  (*localVertexIdentifier)[i].at(
+                    (*localVertexIdentifier)[i].size() - 2),
+                  (*seedIdentifier)[i], (*forkIdentifier)[i],
+                  globalCellId.at(startCellId - 1), startVertexId,
+                  startCellId - 1});
+            } else {
+              globalVertexId.at(startVertexId) = vertIndex;
+              vertIndex++;
+            }
+            startVertexId++;
+          }
+        } else {
+          break;
+        }
+      }
+      seedIdentifier++;
+      forkIdentifier++;
+      localVertexIdentifier++;
+      trajectory++;
+    }
+  }
+  // unmatchedGhosts contains the ghost data for each process.
+  // For two processes i and j that are neighbors, their respective
+  // unmatchedGhosts[k_i] and unmatchedGhosts[k_j] have the same
+  // number of elements and each element represents a segment of
+  // integral line. This means that if unmatchedGhosts[k_i] and
+  // unmatchedGhosts[k_j] are sorted using the same comparator,
+  // element l of unmatchedGhosts[k_i] and element l of unmatchedGhosts[k_j]
+  // represent the same segment of integral line. Therefore, once
+  // unmatchedGhosts vectors are sorted for each process, all that
+  // is left to do is exchange the global identifiers of vertices
+  // and edges, as the receiving process already knows to which local
+  // vertices in its domain the ghosts corresponds to.
+  if(ttk::MPIsize_ > 1) {
+    for(int i = 0; i < neighborNumber_; i++) {
+      TTK_PSORT(threadNumber_, unmatchedGhosts.at(i).begin(),
+                unmatchedGhosts.at(i).end());
+    }
+    this->exchangeGhosts(unmatchedGhosts, globalVertexId, globalCellId);
+  }
+
+  return 0;
+}
+
+int ttk::IntegralLines::exchangeGhosts(
+  std::vector<std::vector<ttk::intgl::GhostElementsToSort>> &unmatchedGhosts,
+  std::vector<ttk::SimplexId> &globalVertexId,
+  std::vector<ttk::SimplexId> &globalCellId) {
+  ttk::SimplexId id{0};
+  std::vector<std::vector<ttk::SimplexId>> globalIdsToSend(neighborNumber_);
+  std::vector<std::vector<ttk::SimplexId>> globalIdsToReceive(neighborNumber_);
+  for(int i = 0; i < neighborNumber_; i++) {
+    globalIdsToSend[i].resize(2 * unmatchedGhosts[i].size());
+    globalIdsToReceive[i].resize(2 * unmatchedGhosts[i].size());
+  }
+  // The sending buffer is prepared
+#ifdef TTK_ENABLE_OPENMP
+#pragma omp parallel for
+#endif
+  for(int i = 0; i < neighborNumber_; i++) {
+    for(size_t j = 0; j < unmatchedGhosts[i].size(); j++) {
+      globalIdsToSend[i][2 * j] = unmatchedGhosts[i][j].ownedGlobalId;
+      globalIdsToSend[i][2 * j + 1] = unmatchedGhosts[i][j].globalEdgeId;
+    }
+  }
+  // Each process sends and receives global identifiers of ghosts
+  // to its neighbors
+  for(int neighbor : neighbors_) {
+    MPI_Sendrecv(globalIdsToSend[neighborsToId_[neighbor]].data(),
+                 globalIdsToSend[neighborsToId_[neighbor]].size(),
+                 ttk::getMPIType(id), neighbor, ttk::MPIrank_,
+                 globalIdsToReceive[neighborsToId_[neighbor]].data(),
+                 globalIdsToSend[neighborsToId_[neighbor]].size(),
+                 ttk::getMPIType(id), neighbor, neighbor, ttk::MPIcomm_,
+                 MPI_STATUS_IGNORE);
+  }
+  // The global identifiers of ghosts are inserted in the globalVertexId
+  // and globalCellId vectors
+#ifdef TTK_ENABLE_OPENMP
+#pragma omp parallel for
+#endif
+  for(int i = 0; i < neighborNumber_; i++) {
+    for(size_t j = 0; j < unmatchedGhosts[i].size(); j++) {
+      globalVertexId.at(unmatchedGhosts[i][j].ghostVertexLocalId)
+        = globalIdsToReceive[i][2 * j];
+      if(globalCellId.at(unmatchedGhosts[i][j].ghostEdgeLocalId) == -1) {
+        globalCellId.at(unmatchedGhosts[i][j].ghostEdgeLocalId)
+          = globalIdsToReceive[i][j * 2 + 1];
+      }
+    }
+  }
+  return 0;
+}
+
+#endif // TTK_ENABLE_MPI
